requirements.txt
PyQt5
PyQtWebEngine
pyqt5-tools
fbs
jinja2
mysql-connector-python